# Reduce complexity overhead (and avoid one copy) when obtaining character SSDQs in `SpriteTextDrawNode`