--- a/osu.Framework/Graphics/Sprites/SpriteText_DrawNode.cs
+++ b/osu.Framework/Graphics/Sprites/SpriteText_DrawNode.cs
@@ -2,7 +2,8 @@
 // See the LICENCE file in the repository root for full licence text.
 
 using System;
-using System.Collections.Generic;
+using System.Buffers;
+using System.Diagnostics;
 using osu.Framework.Graphics.Colour;
 using osu.Framework.Graphics.Primitives;
 using osu.Framework.Graphics.Rendering;
@@ -22,7 +23,9 @@ namespace osu.Framework.Graphics.Sprites
             private ColourInfo shadowColour;
             private Vector2 shadowOffset;
 
-            private readonly List<ScreenSpaceCharacterPart> parts = new List<ScreenSpaceCharacterPart>();
+            private ScreenSpaceCharacterPart[]? parts;
+
+            private int partCount;
 
             public SpriteTextDrawNode(SpriteText source)
                 : base(source)
@@ -33,8 +36,7 @@ namespace osu.Framework.Graphics.Sprites
             {
                 base.ApplyState();
 
-                parts.Clear();
-                parts.AddRange(Source.screenSpaceCharacters);
+                updateScreenSpaceCharacters();
                 shadow = Source.Shadow;
 
                 if (shadow)
@@ -46,6 +48,8 @@ namespace osu.Framework.Graphics.Sprites
 
             public override void Draw(IRenderer renderer)
             {
+                Debug.Assert(parts != null);
+
                 base.Draw(renderer);
 
                 BindTextureShader(renderer);
@@ -58,7 +62,7 @@ namespace osu.Framework.Graphics.Sprites
                 var finalShadowColour = DrawColourInfo.Colour;
                 finalShadowColour.ApplyChild(shadowColour.MultiplyAlpha(shadowAlpha));
 
-                for (int i = 0; i < parts.Count; i++)
+                for (int i = 0; i < partCount; i++)
                 {
                     if (shadow)
                     {
@@ -77,6 +81,44 @@ namespace osu.Framework.Graphics.Sprites
                 }
 
                 UnbindTextureShader(renderer);
+            }
+
+            /// <summary>
+            /// The characters in screen space. These are ready to be drawn.
+            /// </summary>
+            private void updateScreenSpaceCharacters()
+            {
+                partCount = Source.characters.Count;
+
+                if (parts == null || parts.Length < partCount)
+                {
+                    if (parts != null)
+                        ArrayPool<ScreenSpaceCharacterPart>.Shared.Return(parts);
+                    parts = ArrayPool<ScreenSpaceCharacterPart>.Shared.Rent(partCount);
+                }
+
+                Vector2 inflationAmount = DrawInfo.MatrixInverse.ExtractScale().Xy;
+
+                for (int i = 0; i < partCount; i++)
+                {
+                    var character = Source.characters[i];
+                    parts[i] = new ScreenSpaceCharacterPart
+                    {
+                        DrawQuad = Source.ToScreenSpace(character.DrawRectangle.Inflate(inflationAmount)),
+                        InflationPercentage = new Vector2(
+                            character.DrawRectangle.Size.X == 0 ? 0 : inflationAmount.X / character.DrawRectangle.Size.X,
+                            character.DrawRectangle.Size.Y == 0 ? 0 : inflationAmount.Y / character.DrawRectangle.Size.Y),
+                        Texture = character.Texture
+                    };
+                }
+            }
+
+            protected override void Dispose(bool isDisposing)
+            {
+                base.Dispose(isDisposing);
+
+                if (parts != null)
+                    ArrayPool<ScreenSpaceCharacterPart>.Shared.Return(parts);
             }
         }