Of note, this doesn't show a huge improvement in benchmarks, but in an edge case noticed in osu!, the `AddRange` operation can add a perceivable overhead:

![JetBrains Rider 2023-07-03 at 08 17 17](https://github.com/ppy/osu-framework/assets/191335/d4afac1c-1209-4ce0-a404-c665675fe436)

| Before | After |
| :---: | :---: |
| ![JetBrains Rider 2023-07-03 at 08 16 07](https://github.com/ppy/osu-framework/assets/191335/c20539ce-9cab-43a4-8cbd-2ee16fcbfad2) | ![JetBrains Rider 2023-07-03 at 08 15 11](https://github.com/ppy/osu-framework/assets/191335/5213b2e7-a2a6-417c-b4d3-04102569b4fb) |

We believe this may be a dotnet runtime quirk.

Even though this change can't be shown in isolated benchmarks, I'd argue the code quality improvement is worth it.

Of note, the caching of the SSDQs at `SpriteText` was a bit redundant as it was being invalidated by basically everything. So it makes sense to just fetch it every time, regardless, in the draw node itself.

The only potential saving we could obtain with the previous logic would be to shift the load to the `Update` frame. But this wasn't being done. Rather than investigating whether that has any benefits, I'd rather focus on getting `SpriteText` rewritten to use an optimised shader.

master

|           Method |        Mean |     Error |    StdDev | Allocated |
|----------------- |------------:|----------:|----------:|----------:|
|   TestStaticText |    984.4 us |   1.30 us |   1.22 us |       6 B |
|   TestMovingText |  2,701.5 us |  47.68 us |  42.27 us |     117 B |
| TestChangingText | 11,381.6 us | 161.87 us | 143.49 us |   64182 B |

this pr (pooled at 2967626edae655123d64ac5a8ef10d683ddece38)

|           Method |        Mean |     Error |    StdDev | Allocated |
|----------------- |------------:|----------:|----------:|----------:|
|   TestStaticText |    985.3 us |   2.46 us |   2.30 us |       6 B |
|   TestMovingText |  2,928.0 us |  57.80 us |  75.16 us |     119 B |
| TestChangingText | 11,426.4 us | 186.17 us | 155.46 us |   64184 B |


this pr (list at 90a2e286bc84e1bacbd76d8f32606ce7b09685de)

|           Method |        Mean |     Error |    StdDev | Allocated |
|----------------- |------------:|----------:|----------:|----------:|
|   TestStaticText |    985.5 us |   1.80 us |   1.68 us |       6 B |
|   TestMovingText |  2,476.1 us |  47.54 us |  50.86 us |     110 B |
| TestChangingText | 10,793.3 us | 211.44 us | 217.14 us |   64180 B |