@peppy 2023-07-03
> Would like @smoogipoo's take on the invalidation stuff if possible.

I double-checked with him IRL on the invalidation part when making this change, so he should be on board with it 😄