@peppy 2023-07-04
With `EnsureCapacity` (wouldn't expect a change, so just confirming it doesn't get worse):

|           Method |        Mean |    Error |   StdDev | Allocated |
|----------------- |------------:|---------:|---------:|----------:|
|   TestStaticText |    985.8 us |  0.99 us |  0.93 us |       6 B |
|   TestMovingText |  2,529.1 us | 49.84 us | 48.95 us |     108 B |
| TestChangingText | 10,811.6 us | 79.74 us | 74.59 us |   64166 B |